VAVDA: Move Decode job submission onto the GPU ChildProcess.
This moves submission of the decode job (uploading parameters and slices
to the GPU) onto the GPU process ChildThread. This is to work around buggy
Intel libva driver, which appears to have race problems if job submission
for the same VA context happens from separate thread.

This move will have a very minimal effect on performance, it changes the flow
from:

(Decoder thread)
SubmitDecodeBuffers()
ScheduleDecodeJob()
(GPU thread)
Sync()

to doing everything above on the GPU thread. This is not a big problem, as
the operations moved from decoder thread are nonblocking and fast, Sync()
being the heavy one.

BUG=151664
TEST=manual runs


Review URL: https://chromiumcodereview.appspot.com/11026072