Revert gclient --reset for the master
Specifically, reversion of r159593; that CL we want,
but it's proving incompatible w/ the current state of the
builders- specifically, not a large enough svn:ignore in
place to suppress it from screwing w/ those directories.

Thus back out that fix (since it's fixed the immediate issue),
loop back and redo the disk layout so we don't need those ignores,
and then reland r159593.

BUG=None
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/11017020