Tab content flashes grey bottom half after navigating from search
Fixes the flash of grey as contents get reparented from NTP when navigating
to destination page.  Keeps the web_contents detached from the webview while reparenting
and layout is happening.  Except for the case where the contents are loading, then the
web_contents need to be attached but hidden.

BUG=148728
TEST=Manual, as per bug.
R=sky@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11088008