Componentize IgnoreNavigationResourceThrottle and add chrome and webview specific implementations.
This change moves the IgnoreNavigationResourceThrottle to a separate component
as it will be shared between chrome/android and android_webview.
This change also adds the chrome/ and android_webview/ specializations of the
throttle (or more precisely the callback that is used to find the right method
on the right delegate).

BUG=130006
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10946008