cros: Fix window rotation animation
If a user presses Ctrl-Alt-Shift-F3 repeated their windows can get stuck in
a partially rotated state. The rotation animation chooses its target rotation
and transform based on the values at the start of the animation. By default,
if we trigger a new animation it immediately computes a new target based on
the current values.  Instead, queue the animations so we're sure things are
back to neutral before starting a new one.

BUG=154875
TEST=manual, hit Ctrl-Alt-Shift-F3 repeatedly and ensure windows end up in normal position

Review URL: https://chromiumcodereview.appspot.com/11066085