[Mac] Enumerate webcam devices on the main thread.
The associated bug is about a crash in code which is only loaded due
to +deviceNames, though the crash does not have any such code on the
stack.  If that call were not thread-safe, then possibly using it on a
different thread breaks assumptions about where things are allocated
and released.

BUG=139164


Review URL: https://chromiumcodereview.appspot.com/11017045