Add remaining CQ hooks for triggered tests
CQ will now wait for externally submitted tryjobs to complete (pass or fail) before
deciding which steps are still needed on a trybot type.  This is needed for triggered
builders (which do not come from CQ) but should also reduce the number of tryjobs
sent by CQ on external patches, if a tryjob is running when CQ is triggered.

BUG=143485

R=maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11047031