Fix multi-solution use case with conflicting custom_deps overrides.
Let's say there's 2 solutions, solA and solB, both fetching the dependency depA
with different urls, solA has 'depA': 'http://foo/bar' and solB has 'depA':
From('depB', 'bar').

This case used to not work, because LateOverride() was not called in
verify_validity(), so it was not comparing the resolved urls but instead what is
textually specified in both DEPS files.

Also because LateOverride() was not called, verify_validity() was comparing the
original urls even if custom_deps is specified in the solution(s).

Finally, allow one to mismatch the other IFF one is None.

R=szager@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11088023