Disable PluginVarTrackerTest.RecursiveTrackWithNoRef to try to fix ppapi_unittests on win trybot
ppapi_unittests is failing fairly regularly on the windows trybot. The tests are sharded and it's unclear to me exactly where it's failing but it seems to time out when running PluginVarTrackerTest.RecursiveTrackWithNoRef. I've disabled it for now to see if that resolves the problem. See http://build.chromium.org/p/tryserver.chromium/builders/win_rel/builds/69161/steps/ppapi_unittests/logs/stdio for an example of the failure.

BUG=155175
TBR=dmichael,yzshen


Review URL: https://chromiumcodereview.appspot.com/11091055