Reland r161256: Fix the crash that was happening on the buildbot (but not trybots, for a strange reason) because ResourceBundle was initialized twice for tests that use the RenderViewTest harness.
The one difference is to check if ResourceBundle is initialized before initializing it. This is needed because on content_browsertests it's not initialized while in browser_tests it is. We can't make the two browser test binaries behave the same because of differences in how chrome's browser_tests and unit_tests can share the same test suite while in content that's not possible.

TBR=cpu
Review URL: https://codereview.chromium.org/11091058