touch: Cleanup touch-event mode switching on windows.
On windows, chrome used to switch between raw-touch and gesture modes depending on
whether the touch-start event was processed by the renderer or not. But now, the
switch happens iff touch-events is enabled (using command-line flag) on appropriate
version of windows. So cleanup the relevant code.

BUG=138153
TBR=avi@chromium.org (content/port/)

Review URL: https://codereview.chromium.org/11028144