Handle mouse down on the Instant preview.
It turns out that at least on some Windows machines, clicking the
Instant preview doesn't send a PointerActivate event. A MouseDown event
is sent however, so this fixes the bug.

I am not sure why this was never part of the original implementation.
Are there good reasons for not trapping the mouse down event?

BUG=154799
R=sky@chromium.org
TEST=See bug.


Review URL: https://chromiumcodereview.appspot.com/11085095