Fix the android build.
Due to the way make rules are generated from gyp files, the
android_webview rule was being invoked even though we already
had a strip and install action.
The simplest fix is to remove the android_webview rule as we don't
use it anymore.

TBR=torne@chromium.org

BUG=

Review URL: https://codereview.chromium.org/11109014