Change DEPS to include cacheinvalidation code for Android.
The Android build requires more of the cacheinvalidation library than
other ports of Chrome. This changes to use the new full mirror of the
SVN repo that points to trunk/src instead of trunk/src/google.

To simplify transition and to clean up, the folder files is also
removed from the path, and instead we checkout directly to
cacheinvalidation/src.

BUG=154703


Review URL: https://chromiumcodereview.appspot.com/11085063