Merge 161406 - Disable the take-over of XP-era registration on Windows 8.
Eventually, we need to detect whether or not Chrome was made the default handler and take the XP-era registration if so.  That requires some code motion, so for now do nothing so that we at least don't inadvertantly harm our users.

BUG=154806
TEST=go through 1st run of Chrome on Windows 8, choose IE from the "How do you want to open webpages?" dialog, then confirm that HKCU\Software\Classes\.html doesn't point to ChromeHTML.<anything>.

Review URL: https://chromiumcodereview.appspot.com/11094081

TBR=grt@chromium.org
Review URL: https://codereview.chromium.org/11148028