Added backgrounds to hide dead pixels on OS X caused by issue 151985.
This is a temporary hack that just hides the one column of pixels on the
right edge of the calculator window caused by http://crbug.com/154977.
Since Calculator is a v2 app it isn't yet available to regular users on
OS X and only a few developers will see it. This hack should make things
not too ugly for those developers until the underlying problem is fixed.
When that happens, Calculator will find itself fixed without any
additional changes being required, but code that will then be
unnecessary can be removed as described in http://crbug.com/155383.

BUG=154977


Review URL: https://chromiumcodereview.appspot.com/11091078