Revert 162017 - Merge 161036 - Fix a crash on Google Drive operations cancelled during reauthenticaiton.
(It seems I made some mistake during merging, the patch is only partially
merged. I'll revert and retry later.)

This patch introduces a weak pointer to safely notify to the authenticater
that an operations is terminated by the user.

The operations are terminated for three reasons: (1) user cancel, (2)
succesful or failure) completion, and (3) an authentification failure,
later separated from (2).
 - (1) immediately shuts down (2) by URLFetcher::Cancel().
 - (2) and (3) immediately unrolls the operation so (1) won't happen.
 - The steps (2) and (3) are never simultaneously run, so no problem.
So, the only problematic case was the event (1) followed by (3), in
that case a deleted operation object was used.

BUG=153003


Review URL: https://chromiumcodereview.appspot.com/11032031

TBR=kinaba@chromium.org

TBR=kinaba@chromium.org
Review URL: https://codereview.chromium.org/11148033