fix up example extensions options page
The example options.js attempts to attach a listener to the save
button, but at the point the .js gets loaded, that button doesn't
exist, so you end up with a runtime error.  So move the loading
to the end of the example html.

The example options.js also attempts to add an init function, but
that never gets run.  Changing the event to DOMContentLoaded allows
the code to actually work, so change to that instead.

Tested with Chrome Version 24.0.1290.1 dev.

Review URL: https://chromiumcodereview.appspot.com/11151012