Merge 161952 - Prevent platform apps from being able to use the extension tabs API.
r154730 made it not require a permission (except for access to titles and URLs),
but platform apps should not have access to the tabs API in the first place.

Short-term fix is to blacklis the "tabs" and "windows" features for platform
apps.

BUG=155720
R=asargent@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11144019

TBR=mihaip@chromium.org
Review URL: https://codereview.chromium.org/11183003