sync: Merge {Verify,Process}UpdatesCommand
The VerifyUpdatesCommand was used to ensure the validity of each update
that had been delivered from the server and to prepare for the
ProcessUpdatesCommand.  The ProcessUpdatesCommand would then take all
updates that had been successfully verified and move them into the
SERVER_ fields of the sync directory.

It turns out that the logic can be greatly simplified by performing both
tasks within the same command.  This patch does not take full advantage
of the merge.  This patch is intended simply to merge the two files,
with the goal of performing more significant refactorings later.

BUG=154654


Review URL: https://chromiumcodereview.appspot.com/11091009