Add nacl_deps_bump.py, a script for updating nacl_revision in Chromium's DEPS
This tool helps with updating nacl_revision in Chromium's DEPS file to
the latest revision of NaCl.  It creates a Rietveld code review for
the update, listing the new NaCl commits.  It kicks off a try job,
using relevant trybots.

I have been using this tool since July 2011; it has lived in my
"nacl-dev-tools" Git repo on Github.

I'm adding this tool to the 'parasite' directory because I intend to
add tools on top of it which use the tryjob_query.py script (which is
already in the 'parasite' directory).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2862
TEST=none

Review URL: https://chromiumcodereview.appspot.com/11140027