Add --disable-drive-prefetch flag for ChromeOS.
Code for using this flag will follow separately.

The feature (prefetching in Google Drive client for ChromeOS)
is going to be enabled by default, and it won't be user-configurable,
but, for development/debugging purpose it would be good if
we could sometimes turn it off.

BUG=156006

Review URL: https://chromiumcodereview.appspot.com/11151036