Android WebView: make sure libwebview target isn't mangled.
The android gyp backend has been updated to introduce an
"android_unmangled_name" target option, which allows the backend to skip
mangling the name of designated targets. Turn this on for libwebview in
order that it's generated as libwebview.so instead of
lib_android_webview_libwebview_gyp.so which was rather confusing.

BUG=

Review URL: https://codereview.chromium.org/11198030