Merge 161759 - OVDA: Perform an EGL sync before reusing a texture.
When a texture is returned via ReusePictureBuffer(), we know
that it has been inserted into command stream, but it may not
have been read out yet.

This inserts an EGL sync object into the GPU command stream.
All commands that have already been in the stream at the time
of insertion are finished when the object is signaled,
thus guaranteeing that the commands to read out the texture
have been finished. The texture will be reused for decoding
after the object has been signaled.

BUG=155602

Review URL: https://chromiumcodereview.appspot.com/11076009

TBR=posciak@chromium.org
Review URL: https://codereview.chromium.org/11185030