update_nacl_manifest improvements.
Add ability to leave size and checksum blank in the
manifest and have it be filled in based by downloading
the current files.

Add -v/--verbose option which prints more information.

Add -d/--debug option which will show backtraces for
Errors() which is not normally useful.

When in --dryrun mode write the generated file locally
along with the current online one and show a diff of 
the two. 

BUG=155906

Review URL: https://chromiumcodereview.appspot.com/11200002