Fix getting tap events on scroll gestures when touchmove is consumed
Regression introduced in r158592.  The GesturePoint::UpdateForScroll call was
having the side-effect of making the InInsideManhattanSquare check for a click
always return true.  This is a quick fix designed to be easily merged back
to M23 - just don't update the notion of the starting position of a scroll
on consumed touchmove events.

BUG=154844


Review URL: https://chromiumcodereview.appspot.com/11147015