Add test for calling Cancel in ResourceThrottle::WillStartRequest
There used to a DCHECK crash when Cancel is called in
ResourceThrottle::WillStartRequest fixed in r160508. The error was
due to the URLRequest::Start is called in ResourceLoader::StartRequest
after URLRequest::Cancel.

Add a test to explicitly check the request is not started if it Cancel
is called in ResourceThrottle::WillStartRequest.

BUG=

Review URL: https://chromiumcodereview.appspot.com/11044016