[content shell] implement testRunner.display
For layout tests, we only generate frames at defined points (when the test finishes, and when e.g. testRunner.display() is invoked). For each frame, we must only paint the union of all invalidated rects since the last frame.

To achieve this, we forward paint related callbacks from the WebWidgetClient interface to the RenderViewObserver.

With that WebKitTestRunner can mimick the exact behavior of DRT.

BUG=111316
TEST=repaint tests pass

Review URL: https://codereview.chromium.org/11187020