Move pnacl_component_crx_gen to the chrome tree from nacl.
Since it will be used in the chrome tree
for setting up pnacl browser tests, and otherwise isn't
really used in the nacl tree it makes more sense for it
to be in chrome.  This should also make it easier to
coordinate changes between this and pnacl_coordinator in
the chrome tree.

This is almost an exact copy, except for the logic of
finding the PNaCl toolchain.  Since it is in a different
directory now, the search for the toolchain is different.


BUG= none


Review URL: https://chromiumcodereview.appspot.com/11148002