Fixed buttons-are-not-clickable regression introduced in 162476.
Buttons were not being connected to their onclick handler because one of
their classes had changed. Also, buttons were not being identified
correctly because that identification was also based on their class, and
the format of that class changed. Adjusted the selector in the code that
connected buttons to their onclick handler and replaced the error-prone
identification-by-class mechanism with a more resilient identification
by dataset (data-*) attribute.

BUG=156474


Review URL: https://chromiumcodereview.appspot.com/11191046