Move IT2Me functions into their own implementation object.
The script object itself is now used exclusively on the plugin/UI thread.
When the IT2Me host function is used, an internal ref-counted implementation
object is created which runs (mostly) on the network thread. The IT2Me
implementation is passed a weak pointer to the script object through which
to post tasks to it, allowing it to safely out-live the script object, for
instance if there are still pending tasks for it at the moment it is stopped.

BUG=156257
TEST=PyAuto & manual testing of IT2Me host functions.


Review URL: https://chromiumcodereview.appspot.com/11195031