drive: Don't create DriveSystemService if Drive is disabled
This is a reland of crrev.com/162725, which was reverted as it broke
BrowserGuestSessionNavigatorTest.Disposition_Settings_UseIncognitoWindow
on Debug build. FWIW, this wasn't caught by linux_chromeos try bot as
the try bot uses Release build. :~(

There is no need to create DriveSystemService if Drive is disabled.
Note that DriveSystemServiceFactory::GetForProfile() returns NULL
for incognito window and guest login, so the behavior of returning
NULL is consistent with these cases.

BUG=154144,156653
TEST=Drive works as before. Drive is disabled if --disable-gdata is specified.

Review URL: https://codereview.chromium.org/11188073