Make plug-in text fields open the win 8 on-screen keyboard on touch.
This broke when we changed the algorithm used to determine whether the on-screen
keyboard should open on Win 8 - it now relies on having an accurate bounds rectangle
for the current focused text control. This isn't available for plug-ins, so we just
use the whole document size as a quick fix.

This will cause any touch event in a plug-in that causes a text field in the plug-in
to get focus to open the OSK. However, it's possible that touching outside of the
text field could cause the OSK to appear - see http://crbug.com/140993 for an example
where this has happened.

A better fix would be to get the bounds rect of the focused control from plug-ins,
but this is better than nothing for now.

Tested with PDF.

BUG=156534


Review URL: https://chromiumcodereview.appspot.com/11195047