Fix an issue where odd-sized windows would have their bounds rounded up on Mac.
Specifying that an app window should have a size of 359x359 caused the created
window to be 360x360 if an x and y coordinate was not specified, because the x
and y coordinates were being calculated as half the width of the screen in
floating point (i.e. they would end up being fractional).

BUG=151985
R=dharcourt@chromium.org,avi@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11185071