cc: Remove all but one texture uploader class.
This reduces the number of texture uploader classes from 4 to 1.

- TextureUploader
- ThrottledTextureUploader
- UnthrottledTextureUploader
- LayerTextureSubImage

becomes

- TextureUploader

and it lives behind the CCResourceProvider interface where
the LayerTextureSubImage instance used to be.

This also makes the call stack when performing a texture upload less
awkward. It used to look like this:

CCTextureUpdateController::updateMoreTexturesNow() ->
  ThrottledTextureUploader::updateTexture() ->
    CCPrioritizedTexture::upload() ->
      CCResourceProvider::upload() ->
        LayerTextureSubImage::upload() ->
          glTexSubImage2D()

but now looks like this:

CCTextureUpdateController::updateMoreTexturesNow() ->
  CCPrioritizedTexture::upload() ->
    CCResourceProvider::upload() ->
      TextureUploader::upload() ->
        glTexSubImage2D()

which makes a lot more sense considering that
CCTextureUpdateController::updateMoreTexturesNow() should really be called
CCResourceUpdateController::updateMoreResourcesNow() and
CCPrioritizedTexture::upload() should be CCPrioritizedResource::update().

This is only refactoring and should not change the behavior of the
compositor in any way.

BUG=
TEST=cc_unittests


Review URL: https://chromiumcodereview.appspot.com/11188055