cc: Rename viewportPoint to screenSpacePoint to be more precise in the hit-testing code.
The point given to these functions is meant to be applied to the inverse
screenSpaceTransform. That point must be in the device-viewport, ie.
screen space.

The current name makes it sound like the point is in logical pixels rather than
device pixels, which is incorrect.

Just renaming variables, no new tests.

R=enne

Review URL: https://codereview.chromium.org/11191076