Add step seeding and step cursor support to annotator.
This allows us to have more flexible buildsystems on the slave, including ones that operate steps in parallel (swarm). Step seeding allows us to send a manifest of steps we are about to execute, to pre-populate the buildbot waterfall. A further extension would allow seeded steps to be marked tree-closing, providing an annotated solution to gatekeeper. Step cursor allows us to specify which step the subsequent annotator commands apply to, which allow us to take advantage of step seeding and also enable parallel execution. Previous code that relies on BUILD_STEP should be unaffected.

This is RFC until this has been tested on a master (right now it passes unit tests).

BUG=152101


Review URL: https://chromiumcodereview.appspot.com/11109007