Android browser compositor: Handle view surface changes.
When the surface goes away (going to the home screen), call
the synchronous SetVisible() API and tell the compositor the
context needs to be recreated. Because setVisibilityCHROMIUM()
is also synchronous, we know that all GL commands have landed,
and we can return from the SurfaceHolder callback, after which
the surface will be deleted.
Inhibit commits (which would trigger the context to be recreated)
until we have a surface again.

BUG=154815


Review URL: https://chromiumcodereview.appspot.com/11197005