Add basic iOS support to chrome_common.gypi
This is not yet used, but is the foundation for bringing up higher-level
chrome/ gyp files.

The long specific exclusion list is approximately where we expect the
list to end up once upstreaming is done, as a useful reference point
even though it's currently redundant.

BUG=156699

Review URL: https://chromiumcodereview.appspot.com/11188061