[Android] Use DevTools frontend bundled with Content Shell
As the content shell can be compiled from any revision of Chromium,
it's not feasible to deploy all possible versions of DevTools frontend
onto the chrome-devtools.frontend host.

We bundle DevTools frontend with Content Shell, we only need to point the handler to it.

R=satish@chromium.org,pfeldman@chromium.org
BUG=149441

Review URL: https://chromiumcodereview.appspot.com/11190020