cc: Move contentsSwizzled property from LayerTextureUpdater to CCPrioritizedTexture class.
The contentsSwizzled property is true when we upload BGRA contents to
a RGBA texture. Instead of having the LayerTextureUpdater (which is
not responsible for texture uploads) determine the value of this property
set it in CCPrioritizedTexture::upload() when an upload takes place.
This will also allow us to remove the
FrameBufferSkPictureCanvasLayerTextureUpdater class, which currently
only exists so contentsSwizzled is set correctly.

BUG=
TEST=manual

Review URL: https://codereview.chromium.org/11232028