Mac Web Intents: Fix extension install prompt padding
When the extension install prompt was embedded in the web intents dialog it had an additional padding around it.

The problme was that the web intent layout code assummed that all views inside the dialog should have the same padding. This wasn't true for the extension install prompt which has a padding set in the xib file.

To fix this I'm changing the base view controller class to be more flexible. The extension install prompt now overrides -[WebIntentViewController sizeToFitAndLayout] and simply resizes the web intents view to exactly fit the extension install prompt.

Screenshots:
  before: http://i.imgur.com/SmSlA.png
  after: http://i.imgur.com/gXs36.png

BUG=148616
TBR=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11231037