Conditionalize memory allocation limit based on visibility
Replace the single value with a ManagedMemoryPolicy structure
which has the limit in bytes when visible and when not visible,
and also has a priority cutoff when visible and when not visible.

In CCRendererClient, separate the methods for specifying and
enforcing a memory policy. This is so that one may decrease the
consumption of a renderer that has grown over time, without 
disallowing growth in the future.

BUG=134750


Review URL: https://chromiumcodereview.appspot.com/11196038