Merge 162319 - Fix "Please update your sync passphrase" link on Chrome OS settings.
The link was unconditionally associated with browser restart on Chrome OS,
which made users think the browser crashed.
The action was intended for the case when auth error happened and
needs relogin to refresh the authentication credentials when the message
is "Sign out then sign in again".

For the message "Please update your sync passphrase", don't restart the
browser but just show the dialog to prompt sync passphrase.

Instead of assigning a fixed handler for the element ($(sync-action-link)),
change the action according to whether the message is caused by auth error
or anything else.

BUG=150740
TEST=repro the issue described in 150740, open chrome://settings and click on "Please update your sync passphrase" message.

Review URL: https://codereview.chromium.org/11178004

TBR=kochi@chromium.org
Review URL: https://codereview.chromium.org/11234051