Fixed 31 / 3 failure in Calculator default app.
A problem with the rounding logic was causing any operation whose result
was a decimal larger than 10 with more than 7 digits after the decimal
to fail. (31 / 3) is one such operation. This was fixed and a test was
added for this case. Comments were also added to the tests to describing
some other test cases that need to be tested for and/or fixed. The
simplest fix to the rounding problem involved adding support for the
display of exponential values, which will probably fix a couple of
aspects of http://crbug/156453.

BUG=156448
BUG=156453


Review URL: https://chromiumcodereview.appspot.com/11183051