Don't call app.window.create callback until after DidCreateDocumentElement.
This means that in the callback of chrome.app.window.create, the contentWindow.document element will be the final target document, not the initial 'about:blank' document. The callback will still be guaranteed to be called before the child window's onload event happens, and before any scripts run in the child.


R=mihaip@chromium.org,darin@chromium.org
BUG=148857,130465

Review URL: https://chromiumcodereview.appspot.com/10979085