desktop linux aura: Fix event/capture focus after dragging a window.
Both linux_aura versions have had a persistent problem where if you drag a tab
out of a window, you have to click on either window to get mouse input working
in the web content area. This is because the RootWindow still thinks a mouse
drag is in progress.

BUG=146077


Review URL: https://chromiumcodereview.appspot.com/11189148