Add aw test to check UA is actually set by ContentSettings.
The magic code that this is currently testing is at the end of
RenderViewImpl::willSendRequest, which sets the UA override (if there is
one). The network stack will only ask UrlRequestContext for a user agent
if it is not already set.

BUG=
Java test change. Ran through android java trybot
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11227059