Merge 163397 - Support Help and Fn keys on Stout.
The scancodes for these keys are <FK14> and <FK15> respectively, that are translated to XF86XK_Launch5 and XF86XK_Launch6 XKeySyms. We have to convert the XKeySyms to valid VKEYs so that they are not filtered out as invalid keys.

BUG=chrome-os-partner:14339
TEST=manually with an old Apple keyboard which has F13-19 keys

Review URL: https://codereview.chromium.org/11238039

TBR=yusukes@chromium.org
Review URL: https://codereview.chromium.org/11271013