Remove disable_nacl=1 for now.
Otherwise, libppGoogleNaClPluginChrome.so is not built and it is necessary for
sync_integration_tests.

This raises the issue of GYP_DEFINES causing different dependencies to be
required. The .isolate format supports this in theory but it is currently
explicitly not supported, since the list of accepted conditions is done with a
whitelist of variables.

TBR=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11276016