Linux: don't always launch sandbox renderer tests.
If the sandbox has not been initialized, we should not launch the renderer
sandbox tests. This can happen in only one case: if the Zygote is not used and
the sandbox is de-activated.

We make RunSandboxTests() take an extra no_sandbox parameter to match
InitSandboxTests.

BUG=157470
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11175074