Report memory usage statistics from the renderer.
Add functions to PrioritizedTextureManager to report the amount of visible and nearly-visible memory, based on the requested set of textures to draw (note that this is a superset of the backings that will be visible and nearly-visible). Add a unittest for this functionality.

Add infrastructure in the Proxy to send this data to the renderer. From here the data will go to the WebGraphicsContext3D, then the GPU command buffer, and finally the GPU memory manager.

BUG=134750


Review URL: https://chromiumcodereview.appspot.com/11230064