chrome.experimental.identity.launchWebAuthFlow doesn't catch JS redirects
The problem is that WebAuthFlow doesn't get any of the notifications
it listens to when a URL is changed by Javascript.

After trying a few things, looks like WebContentsObserver::ProvisionalChangeToMainFrameUrl
is called when window.location is changed in JS.

Since now WebAuthFlow is a WebContentsObserver, move to using observer callbacks for most
things and stop listening to NOTIFICATION_LOAD_START, NOTIFICATION_LOAD_STOP, NOTIFICATOIN_WEB_CONTENTS_DESTROYED.

It is preferable to use observer anyway since some notifications don't work great. For example, load_details is
always NULL when NOTIFICATION_LOAD_START is called.

Note that we cannot still stop being a NotificationsObserver since in some redirect patterns only the
NOTIFICATION_RESOURCE_RECEIVED_REDIRECT is called. But I reduced the use of notifications to a minimum now.

BUG=148607
TEST=Apart from unit tests, tested with Google, Facebook and the test auth app attached in the bug extensively
in interactive and non-interactive mode, in sign-in and not signed-in state, in pre-approved and not pre-approved state.

Review URL: https://chromiumcodereview.appspot.com/11184055