[Mac] Fix dictionary popup in text areas.
In the case of selected text in a text area, |selection_text_| actually
stores the whole text in the text area (or at least a large range of it),
and the subset that's selected is specified via |selection_text_offset_|
and |selection_range_|. This change fixes the logic when comparing the
text to actually compare that substring instead of the complete range.

This was broken by http://crrev.com/158988.

Note: It mirrors the substring logic done by RenderWidgetHostViewMac::SelectionChanged().

BUG=154608
TEST=see bug instructions
Review URL: https://codereview.chromium.org/11269053