Merge 163884 - Limit extension install cancellation on quit to Chrome OS
We want to backport the extension install cancellation code to Chrome OS for M23, but it's late in the game for Windows/Mac/Linux M23.  Limit the code to Chrome OS, then backport the original change and this change to M23.  Then we can make it cross-platform in trunk for M24.

BUG=155994,157609
TEST=compiles

Review URL: https://codereview.chromium.org/11277003

TBR=jamescook@chromium.org
Review URL: https://codereview.chromium.org/11266059