Roll third_party/cacheinvalidation from r221:r242.
The intention of this roll is to add 2 source types in
proto (COSMO_CHANGELOG and CHROME_PUSH_MESSAGING), but
also includes the v2 of the Android client library.

r241 had an Android build issue (a symbol "ANDROID" was
defined as an enum member in proto, which was blown away
by -DANDROID compiler option, introduced at r239), and
r242 reverted it.

BUG=150005, 157546
TEST=compile succeeds.

Review URL: https://codereview.chromium.org/11194078