sync: Merge apply updates and resolve conflicts
The conflict resolution code was executed after the commit for reasons
which no longer apply.  Because we no longer have to worry about
resolving hierarchy conflicts or nigori node conflicts, we have the
opportunity to move conflict resolution closer to update application.

One advantage of resolving conflicts early is that we no longer require
an extra sync cycle to commit any 'local wins' conflict resolutions.
This makes SYNC_CYCLE_CONTINUATION sync cycles obsolete.

Another advantage is that update application and conflict resolution can
be performed without releasing the sync lock, which eliminates several
types of races that we used to have to worry about.  It's probably more
efficient, too.

It allows us to guarantee that there are no simple conflicts remaining
after the update application step is completed.  The effects might not
be very noticeable to end users, but it will be nice to remove some of
the conflict tracking code.

Finally, it removes the last use PerModelSafeGroupState.  This will let
us delete some unused code and hopefully simplify StatusController.

This patch does not pursue these cleanups as agressively as it could.
The idea here is to keep the complex logic changes in one small CL, and
perform the cleanups in a larger, but simpler, follow-up CL.

BUG=147681,11280,156238


Review URL: https://chromiumcodereview.appspot.com/11192071