Sync test_policy for DriveDisabled with policy_templates.json
The test_policy for DriveDisabled had to be changed to |false| as
attempting to disable drive would crash the browser on shutdown. Now
that the fix for crbug.com/157611 has landed, disabling drive no longer
crashes. The test_policy can be changed back to |true| so that it is
nicely in sync with the example policy value in policy_templates.json.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11304006