Updated Calculator.app manifest.json to match version in Web Store
Calculator was uploaded to the Chrome Web Store with a 1.3 version, so
this change updates the manifest.json file of the checked in Calculator
source to have it match that version number. This change also includes
a removal from the manifest of the offline_enabled: true (true is now
the default value for that property).


Review URL: https://chromiumcodereview.appspot.com/11275013