Disambiguate process ids on Windows
Windows reuses process ids very often and that confuses the thunk process
detection logic. Now look at the command line to make sure the pid noted is
indeed the thunk, by looking at the command line to make sure the script name
and the trace name fits.

Clarify pid that are used as signals so renamed tracer_pid to thunk_pid since in
reality it's not the pid of the tracer.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11337008