ash: Clean up system background layer.
This adds a SystemBackgroundController class that's
responsible for a ui::Layer that's stacked underneath all
other layers.

I'm also adding an --ash-copy-host-background-at-boot flag
that can be set to instruct the controller to copy the host
window's content into the system background layer at
startup; otherwise, we display the same background color
used by Chrome OS's boot splash screen.

I suspect that there's a small performance hit from copying
(I've seen the XGetImage() take between 25 and 170 ms
depending on hardware), although it seems to be hidden in
the noise of our total boot times.  Regardless, I figure
that we should only turn this on right now for systems that
need it (to hide an even-longer delay where the compositor
is drawing but the login background images hasn't been
loaded).

In any case, once the login background is visible, the
system background layer is reset to display the black
background that's needed for fullscreen animations that
manipulate the desktop background (for the workspace or
power button).

BUG=152309


Review URL: https://chromiumcodereview.appspot.com/11273059