Fix use-after-free and crash in extension install from web intents
WebIntentPickerController is a WebstoreInstaller::Delegate, but can be deleted
before the WebstoreInstaller finishes installing the extension, leading to a
use-after-free and crash.  We're seeing this crash on Chrome OS, perhaps due
to timing differences in extension install.

In other places in the code this lifetime issue is addressed by adding
reference counting to the class implementing WebstoreInstall::Delegate and
manually using AddRef() and Release() around the extension installation,
but in this case the lifetime of WebIntentPickerController is controlled by
WebContents.

Fix by manually invalidate the WebstoreInstaller's delegate_ when the
WebIntentPickerController goes away.

BUG=157767
TEST=added to browser_tests WebIntentPickerControllerBrowserTest, also
manually install extension from web intents on imagemator.com


Review URL: https://chromiumcodereview.appspot.com/11315013