Browser Plugin: Guest focus state should always match BrowserPlugin focus state
In a nutshell, the guest should be focused iff the embedder's BrowserPlugin is focused.

This means that if the BrowserPlugin is focused before initial navigation, the guest needs to get focused upon committing its first navigation.

Furthermore, if a guest crashes and a new guest process is created, the new guest process should inherit the focus state of the previous guest.

BUG=157293
Test=BrowserPluginHostTest.FocusBeforeNavigation, BrowserPluginHostTest.FocusPreservation


Review URL: https://chromiumcodereview.appspot.com/11228043