Fix --enable-stats-table crash on Release builds.
Currently, using the --enable-stats-table flag when running a Release build of chromium causes a segfault on startup.

I traced the issue back to a dangling reference to a local variable that had gone out of scope, so dynamically allocating the variable and storing it as a class member fixes the issue.

BUG=137243


Review URL: https://chromiumcodereview.appspot.com/11260033