Retry the sideload wipeout polish changelist.
This is the same changelist that was reviewed here:
http://codereview.chromium.org/11232060/
... with a few changes:
1) Moved the RemoveObserver call (in ToolbarView) that was messing up interactive tests to happen when we've received the message we are after.
2) Change the code to use UpdatesFromGallery in two places instead of from_webstore.
3) Added FeatureSwitch flag to ExtensionStartupTests, to avoid Sideload Wipeout from affecting the tests.
4) Added a unit test.
5) Make sure sideload-wipeout is disabled by default.

Description from previous changelist (copied unchanged):

Make sure sideload wipeout doesn't interfere with the tests.
Make sure it does not run on other platforms besides Windows, where it was intended (for now).
Fix the problem with bubble appearing to the right of the browser.
Make sure the bubble does not appear on next run after explicitly dismissing it.

BUG=154624
TEST=None
Review URL: https://codereview.chromium.org/11266007