Use proper texture size for texture readback.
If PopulateBitmapWitHContents is called right after a resize, we hit the
glReadPixels path that reads back single lines at a time, which is really
painful on Android devices (10 seconds to readback the content!).  This was
because we were using the set size, not the size the renderer gave us.  This
only happens when we call this method right after resizing as the new texture
with the updated size hasn't been set yet.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11301027