Add "dump to file" button to about:sync.
In the Data tab, there's a "dump to text" button. This adds a
"dump to file" button which triggers a download of a bunch of
debugging data suitable for attaching to a bug.

The format is fake CSV. There's a header with date/time plus
useragent, followed by field labels, and then the data.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11232067