[Mac] Suppress and debug "invalid drawable" messages.
Per the bug, this message happens when -[NSOpenGLContext setView:] is
called with a view that is either not in a window, or who's window
does not have backing store.  This checks for those cases before
continuing, and in those cases records breakpad keys which will
hopefully point to a repro case.

BUG=148882
TEST=none


Review URL: https://chromiumcodereview.appspot.com/11275064