Reserve resource ids starting at 31000 for projects built on Chromium.
Currently iOS has a bunch of strings in the internal repository and they are
repacked by the build in the same .pak. The intent of this CL is to have these 
strings in a protected range to prevent the possibility of sudden build
bustage due to collisions. As this seems like a general problem that anything
built on Chromium is likely to face, this CLs reserves the resource ids starting at
31000 for all projects built on top of Chromium.

Review URL: https://chromiumcodereview.appspot.com/11341009