Switch the TCP reads on Windows to use non-blocking/non-async I/O. The Overlapped I/O was introducing delays when the networking stack did not have enough data to fill the receive buffer.
This can be seen when loading pssplayground.com/ksimbili/webp.html
using a DSL connection profile on WebPagetest:
http://www.webpagetest.org/result/120830_MS_414849a6aa055bb853e7e5d51e1b29d8/
and manifests and increasingly long Time to First Byte for requests
further down the waterfall (expected values are < 90ms and it was 
going over 150ms).

It is configured as a 50% field trial and can be forced through the 
command-line for testing:
    --overlapped-reads=on - default/existing behavior
    --overlapped-reads=off - new read implementation

Trial-specific histograms are reported for page load times and
http request times.  Specifically:
    PLT.Abandoned
    PLT.LoadType
    PLT.BeginToFinish_NormalLoad
    PLT.BeginToFinish_LinkLoadNormal
    PLT.BeginToFinish_LinkLoadReload
    PLT.BeginToFinish_LinkLoadStaleOk
    Net.HttpJob.TotalTime
    Net.HttpJob.TotalTimeSuccess
    Net.HttpJob.TotalTimeCancel
    Net.HttpJob.TotalTimeCached
    Net.HttpJob.TotalTimeNotCached

Review URL: https://chromiumcodereview.appspot.com/10916016