Add method for checking whether the current account is a public account
This CL adds a method to the UserManager that indicates whether the
currently logged-in account is a public account. Since it is not possible
to log into a public account yet, the method always returns |false| for
now. However, it can be mocked, allowing public account features to be
developed and tested before login support is ready.

BUG=158509


Review URL: https://chromiumcodereview.appspot.com/11341042