sync: add kInitialBackoffImmediateRetry
The primary impact this has is speeding up some integration tests by
~ 50%. In particular it lets us re-enable MigrationHell, which is
currently disabled as it takes too long for sharding_supervisor. 

Rationale is discussed in the bug, I believe this is safe for the class
of errors we target with the immediate delay, because we don't need backoff.

(Note that when used with GetDelay, this still results in a 1s minimum
backoff.)

kInitialBackoffShortRetry is still used if the command line flag is passed
(and in integration tests) to change the default initial backoff from 5 minutes
to 1s (as it used to be).

BUG=143641

Review URL: https://chromiumcodereview.appspot.com/11183073