Added touch support to the Calculator default app.
Previously the calculator supported touch only through the system
default interpret-touch-events-as-mouse-ones. This introduced a 150ms
delay on every key touch as the system paused to make sure the touch
should be interpreted as a click rather than a scroll. This change
add support for touch events to the calculator so it can do the right
thing without a pause.

BUG=155574


Review URL: https://chromiumcodereview.appspot.com/11276043