Fix version discovery for Interfaces, Structs
Now that we are looking at including structures and interfaces pointers in the
version graph, we need to fix remove some simplifying  assumptions, and correct
the versioning calculation.

Includes minor cleanup to how we call 'Main' to make it more compliant as
well as catching exceptions so that we show the commandline options that caused
that exception.

BUG=154127,158980
TBR=sehr@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11263062