Adjust gpu memory manager limits for Android for high-res tablets.
We could return the limit from the Android egl context (since it 
knows its surface size), but it doesn't quite fit with the code 
in UpdateAvailableGpuMemory, and would be coupled to the other 
logic in there, so I decided to keep all the code in one place. 

This also converges behavior and tests. The behavior is the 
same now, minus calculating the total memory available which 
has a new test. Since only foreground tabs/windows are given 
memory this should work fine. However, we will need to be 
careful when enabling partial evictions of background tabs, 
as Android assumes background tabs are fully evicted. 

BUG=158465


Review URL: https://chromiumcodereview.appspot.com/11360047