accelerated_surface_win.cc: Two CopyTo bug fixes, and new event tracing.
Bug 1: It seems that nothing really guarantees that src_subrect argument
of CopyTo is in sync with the actual back buffer size. In practice I
saw some failures here (with the D3D9 debug mode enabled) when
resizing windows -- src_subrect was bigger than the allocated
back buffer size. So I've added clipping.

Bug 2: GetRenderTargetData requires width and height of src and dst
to be exactly identical. There was logic in place to handle this,
but it wasn't quite correct -- for example, if you have a src_subrect
that was smaller than dst_size in both dimensions, GetResampleCount
previously returned 0, resulting in an illegal (failing) call to
GetRenderTargetData.

Tracing: I've added event tracing to the possibly expensive D3D calls inside CopyTo. I also added tracing to show the time spent waiting on the present lock, which in my observation was sometimes a significant blocking event for the UI thread [ :( ].


Review URL: https://chromiumcodereview.appspot.com/11367019