Clean up memory manager's context tracking structures.
This change should not change any behavior.

Use GetMemoryTracker and comparisons instead of IsInSameContextShareGroup to
determine if two stubs should be counted as being in the same group.

Simplify the logic in GpuMemoryManager::Manage() use an O(log n) lookup based
on memory tracker instead of an O(n) lookup based on stub context group to
determine if two stubs are in the same context group.

Fix a number of places in the memory manager unit test where "is in the same
context group as" was assumed not to be an equivalence relation (now that
the lookup is based on memory tracker, it is an equivalence relation).

Add memory tracking to the memory manager unit test (so that we can simulate
unmanaged memory, e.g, WebGL, entering the mix).

Clean up the memory manager unit test to use a common base class for the two
stub classes.

BUG=134750


Review URL: https://chromiumcodereview.appspot.com/11361050