Don't return NULL from AudioManager:GetMessageLoop()
Returning NULL leads to a bunch of hairy corner cases where
other clients on the audio thread need to constantly check if
the message loop is still valid.

Returning NULL from GetMessageLoop seems to be a relic from
the past, so for don't do it anymore.  If we see browser
shutdown hangs in the field we can revert and consider
alternate solutions.

BUG=157957
TEST=all tests still pass.

Review URL: https://chromiumcodereview.appspot.com/11348016