Pepper Flash Mac: Fix focus issue when window activation changes
The Pepper Flash plugin was not getting focus change events for window activation changes. For example, if a browser window was deactivated then ppapi::PluginInstance::SetContentAreaFocus(false) was not being called.

To fix this I'm changing RenderWidgetHostViewMac::SetActive() to call Focus() and Blur() as necessary. This matches the behavior of the Windows code which also does the same thing.

BUG=145008
TEST=Verified that the game in http://armorgames.com/play/10606/elephant-quest correctly pauses when window focus is lost.

Review URL: https://chromiumcodereview.appspot.com/11364017