[Win8] Fix pin / unpin status if the user cancels an action or pins / unpins from start screen.
Previously the state was cached whenever a navigation happened, or whenever the user started a pin / unpin action. As the state query does not happen on the metro thread (it happens on the UI thread and is synchronous) there is no need for the caching.

BUG=144332


Review URL: https://chromiumcodereview.appspot.com/11362050