ui: Add IsExpressibleAsRect() method to RectF
This method reports if the RectF can be safely converted to an
integer Rect. When it is false, some dimension of the RectF is
outside the bounds of what an integer can represent, and
converting it to a Rect will require clamping.

Tests:
ui_unittests:RectTest.IsExpressibleAsRect

TBR=sky
BUG=147395
Relanding: https://codereview.chromium.org/11364054/

Review URL: https://codereview.chromium.org/11359052