File Manager: Change the mount status even when the mount is not requested.
If the mount is requested from C++ side (instead of UI side),  the mount status will not be changed at mountGData() because mountCompleted event is occurred at first and mountGData() is never called. In such a case, the mount status is changed at onMountCompleted() instead of mountGData().

BUG=158022
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/11312002