Added completion notification to the call chain from Profile's ClearNetworkingHistorySince() to HttpServerPropertiesManager's Clear(), so now the "Clear Browsing Data" dialog will not close before clearing for this particular component has been completed.
Updated the unit test HttpServerPropertiesManagerTest.Clear() to expect a callback, and used existing browser tests to make sure that the delegation of the callback all the way from ClearNetworkingHistorySince() works properly.

TBR-ing Jay because of a minor change to TestingProfile, which was required because of a minor change to the Profile interface, which was in turn reviewed by one of the Profile owners.

BUG=151615
TBR=jcivelli@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11238034