BrowserPlugin: Move resize logic to BrowserPluginGuest
I'm working on AutoSize and I need to call the resize code from BrowserPluginGuest.
This patch moves the complexity of BrowserPluginEmbedder::ResizeGuest to
BrowserPluginGuest::Resize which feels appropriate.

This patch is a step towards reducing BrowserPluginEmbedder's role to that of a 
message router and lifetime manager of BrowserPluginGuests which makes plumbing
more obvious.

BUG=none
Test=BrowserPluginHostTest.*, BrowserPluginTest.*


Review URL: https://chromiumcodereview.appspot.com/11367086