Fix UV plane interpolation for YUV formats
Follow-on to r165103.  Turns out that we want smooth interpolation on the
UV planes.  Also removes the unittest added to check that straight conversion
and 1:1 scaling give identical results -- this will fail because straight
conversion is presently buggy and does not do UV interpolation.

BUG=chromium:158462
TEST=local build, run on x86, unittests


Review URL: https://chromiumcodereview.appspot.com/11358022