Android: Report step failure when tests time out.
Currently when tests time out, we retry them but still mark the step as
a warning, which does not cause a step failure.

We should report a failure when tests time out. Also adding more stats
on how many tests we were supposed to run and how many we failed to run.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11363088