Create helper function to determine label for sign-out button
This CL moves the code that determines the label for the sign-out button
shown in the aura bar context menu to a helper function. This will allow
the code to be shared with the upcoming aura bar sign-out button for
public accounts, ensuring the labels on both buttons are always in sync.

BUG=152929


Review URL: https://chromiumcodereview.appspot.com/11338041