Add chromeos::NetworkStateManager to src/chromeos/network.
The intention for this class is to provide up-to-date state information about the network for the majority of the network related UI. It should also serve as a foundation for classes monitoring the state of the active network.

This code is not designed to provide detailed information about specific network service or device properties, not should it provide the ability to set (configure) those properties.

BUG=154072

Review URL: https://chromiumcodereview.appspot.com/11192024