Add a SafeSearch preference, policy and implementation.
When the SafeSearch preference is set and the user searches on Google the
safe=active&ssui=on parameters are added at the end of the query if they are
not already set to the correct value. We do this after the request was handled by
extensions.

Also added a policy tied to the SafeSearch preference and both unit tests for the
SafeSearch code and browser tests for the policy part.

TBR=thakis@chromium.org
BUG=159241
TEST=Included unit and browser tests.

Review URL: https://chromiumcodereview.appspot.com/11186002