Upstream recent downstream changes in envsetup.sh.
A superset of this change was needed to fix a bug downstream. Note that the bug
doesn't exist upstream since it was caused by a piece of code not upstreamed.

Therefore the only purpose of this change is to decrease the upstream vs
downstream diff.


Review URL: https://chromiumcodereview.appspot.com/11365147