Chrome remote control multipage tests: Add interactions.
The page set .json description defines what kind of interactions (e.g., link
clicks) are possible for a page. The corresponding Web Page Replay .wpr file
must contain replies to the requests sent when the interaction is executed.

This will be used in memory tests (see
http://codereview.chromium.org/11273081/ ).

The top_25_usage.json is a proof of concept page set description adding a couple
of interactions. When it's more complete, the existing tests will be migrated to
use it.

Parts of the code are stolen from nduca@chromium.org (see
http://codereview.chromium.org/11114020/ ).

BUG=158323

Review URL: https://codereview.chromium.org/11369075