Adds a header to Gaia request to indicate that Gaia can show the user the chrome sync interstitial.
This CL depends on http://codereview.chromium.org/11148029/
This CL is a dependent of http://codereview.chromium.org/11092029/

BUG=137118
TEST=Go to gmail.com and sign in.  Make sure to use a profile that is not
already connected to a google account.  After signing in, see the interstitial
asking if you want to connect your profile to your google account.  This
is currently only available to a few users for testing.


Review URL: https://chromiumcodereview.appspot.com/11188002