aura-x11: Ignore the release events for gesture buttons.
When one of the gesture-buttons on a mouse (i.e. back/forward buttons) is pressed,
it triggers a user-action (which may do a navigation). For these buttons, the release
events need to be ignored. Otherwise it results in a spurious mouse-up event in
WebCore, which results in a crash.

BUG=159685

Review URL: https://codereview.chromium.org/11368109